+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -18,12 +18,14 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (mut data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) = msgpack::deserialize(reader).unwrap();
+    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+        msgpack::deserialize(reader).unwrap();
+
     for channel in data.channels.iter_mut() {
         channel.chat.events.set_stable_memory_key_prefixes();
     }
 
-    canister_logger::init_with_logs(data.test_mode, logs, traces);
+    canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
     let env = init_env(data.rng_seed);
     init_state(env, data, args.wasm_version);