+++ backend/canisters/storage_index/impl/src/lifecycle/post_upgrade.rs
@@ -16,7 +16,9 @@ fn post_upgrade(args: Args) {
 
     let (data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) = msgpack::deserialize(reader).unwrap();
 
-    canister_logger::init_with_logs(data.test_mode, logs, traces);
+    // TODO: After release change this to
+    // let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) = msgpack::deserialize(reader).unwrap();
+    canister_logger::init_with_logs(data.test_mode, Vec::new(), logs, traces);
 
     let env = init_env(data.rng_seed);
     init_cycles_dispenser_client(