+++ frontend/app/src/components/home/addgroup/NewGroup.svelte
@@ -75,7 +75,8 @@
     $: avatarDirty = editing && candidateGroup.avatar?.blobUrl !== originalGroup.avatar?.blobUrl;
     $: visDirty = editing && candidateGroup.public !== originalGroup.public;
     $: infoDirty = nameDirty || descDirty || avatarDirty || externalUrlDirty;
-    $: gateDirty = editing && client.hasAccessGateChanged(candidateGroup.gate, originalGroup.gate);
+    $: gateDirty =
+        editing && client.hasAccessGateChanged(candidateGroup.gateConfig, originalGroup.gateConfig);
     $: ttlDirty = editing && candidateGroup.eventsTTL !== originalGroup.eventsTTL;
     $: messagesVisibleToNonMembersDirty =
         editing &&
@@ -236,7 +237,7 @@
                         ? "set_to_none"
                         : { value: updatedGroup.eventsTTL }
                     : undefined,
-                gateDirty ? updatedGroup.gate : undefined,
+                gateDirty ? updatedGroup.gateConfig : undefined,
                 visDirty ? updatedGroup.public : undefined,
                 messagesVisibleToNonMembersDirty
                     ? updatedGroup.messagesVisibleToNonMembers
@@ -499,6 +500,8 @@
         height: 550px;
         display: flex;
         @include nice-scrollbar();
+        overflow-y: scroll;
+        scrollbar-gutter: stable;
 
         @include mobile() {
             height: 400px;