+++ backend/canisters/registry/impl/src/updates/add_token.rs
@@ -1,23 +1,30 @@
 use crate::guards::caller_is_governance_principal;
 use crate::metadata_helper::MetadataHelper;
-use crate::mutate_state;
+use crate::{mutate_state, read_state};
+use candid::Principal;
 use canister_api_macros::proposal;
 use canister_tracing_macros::trace;
+use icrc_ledger_types::icrc2::transfer_from::TransferFromArgs;
 use registry_canister::add_token::{Response::*, *};
-use registry_canister::NervousSystemDetails;
+use registry_canister::{NervousSystemDetails, Payment};
 use tracing::{error, info};
-use types::CanisterId;
+use types::{CanisterId, Cryptocurrency, UserId};
+use utils::consts::{MEMO_LIST_TOKEN, SNS_GOVERNANCE_CANISTER_ID};
+
+const TOKEN_LISTING_FEE_E8S: u128 = 50_000_000_000; // 500 CHAT
 
 #[proposal(guard = "caller_is_governance_principal")]
 #[trace]
 async fn add_token(args: Args) -> Response {
     add_token_impl(
         args.ledger_canister_id,
+        args.submitted_by,
         None,
         Some(args.info_url),
         Some(args.how_to_buy_url),
         Some(args.transaction_url_format),
         args.logo,
+        true,
     )
     .await
 }
@@ -25,22 +32,27 @@ async fn add_token(args: Args) -> Response {
 pub(crate) async fn add_sns_token(nervous_system: NervousSystemDetails) {
     add_token_impl(
         nervous_system.ledger_canister_id,
+        None,
         Some(nervous_system),
         None,
         None,
         None,
         None,
+        false,
     )
     .await;
 }
 
+#[allow(clippy::too_many_arguments)]
 async fn add_token_impl(
     ledger_canister_id: CanisterId,
+    submitted_by: Option<UserId>,
     nervous_system: Option<NervousSystemDetails>,
     info_url: Option<String>,
     how_to_buy_url: Option<String>,
     transaction_url_format: Option<String>,
     logo: Option<String>,
+    requires_payment: bool,
 ) -> Response {
     let metadata = match icrc_ledger_canister_c2c_client::icrc1_metadata(ledger_canister_id).await {
         Ok(r) => r,
@@ -68,47 +80,96 @@ async fn add_token_impl(
         }
     };
 
-    if let Some(logo) = metadata_helper
+    let Some(logo) = metadata_helper
         .logo()
         .cloned()
         .or(logo)
         .or(nervous_system.as_ref().map(|ns| ns.logo.clone()))
-    {
-        match icrc_ledger_canister_c2c_client::icrc1_supported_standards(ledger_canister_id).await {
-            Ok(standards) => mutate_state(|state| {
-                let now = state.env.now();
-                let standards = standards.into_iter().map(|r| r.name).collect();
-
-                if let Some(ns) = nervous_system {
-                    state.data.nervous_systems.add(ns, now);
-                }
-                let name = metadata_helper.name().to_string();
-                if state.data.tokens.add(
-                    ledger_canister_id,
-                    name.clone(),
-                    metadata_helper.symbol().to_string(),
-                    metadata_helper.decimals(),
-                    metadata_helper.fee(),
-                    logo,
-                    info_url,
-                    how_to_buy_url,
-                    transaction_url_format,
-                    standards,
-                    now,
-                ) {
-                    info!(name, %ledger_canister_id, "Token added");
-                    Success
-                } else {
-                    AlreadyAdded
-                }
-            }),
-            Err(error) => InternalError(format!("{error:?}")),
-        }
-    } else {
+    else {
         let error = "Failed to find logo for token";
         error!(%ledger_canister_id, error);
-        InternalError(error.to_string())
+        return InternalError(error.to_string());
+    };
+
+    let standards = match icrc_ledger_canister_c2c_client::icrc1_supported_standards(ledger_canister_id).await {
+        Ok(standards) => standards,
+        Err(error) => return InternalError(format!("{error:?}")),
+    };
+
+    let (test_mode, now_nanos, exists) = read_state(|state| {
+        (
+            state.data.test_mode,
+            state.env.now_nanos(),
+            state.data.tokens.exists(ledger_canister_id),
+        )
+    });
+
+    if exists {
+        return AlreadyAdded;
     }
+
+    // BURN the listing fee from given user if required
+    let mut payment: Option<Payment> = None;
+    if requires_payment && !test_mode {
+        if let Some(user_id) = submitted_by {
+            let amount = TOKEN_LISTING_FEE_E8S;
+            let from: Principal = user_id.into();
+            let transfer_args = TransferFromArgs {
+                spender_subaccount: None,
+                from: from.into(),
+                to: SNS_GOVERNANCE_CANISTER_ID.into(),
+                amount: amount.into(),
+                fee: Cryptocurrency::CHAT.fee().map(|fee| fee.into()),
+                memo: Some(MEMO_LIST_TOKEN.to_vec().into()),
+                created_at_time: Some(now_nanos),
+            };
+
+            match icrc2_transfer_from(Cryptocurrency::CHAT.ledger_canister_id().unwrap(), &transfer_args).await {
+                Ok(block_index) => {
+                    payment = Some(Payment {
+                        amount,
+                        block_index,
+                        timestamp: now_nanos / 1000,
+                        user_id,
+                    });
+                }
+                Err(message) => {
+                    error!(%user_id, ?message, "Error transferring listing fee");
+                    return PaymentFailed("message".to_string());
+                }
+            }
+        } else {
+            return InvalidRequest("Request missing `submitted_by` required to pay for token listing".to_string());
+        }
+    }
+
+    mutate_state(|state| {
+        let now = state.env.now();
+        let standards = standards.into_iter().map(|r| r.name).collect();
+
+        if let Some(ns) = nervous_system {
+            state.data.nervous_systems.add(ns, now);
+        }
+
+        let name = metadata_helper.name().to_string();
+        state.data.tokens.add(
+            ledger_canister_id,
+            name.clone(),
+            metadata_helper.symbol().to_string(),
+            metadata_helper.decimals(),
+            metadata_helper.fee(),
+            logo,
+            info_url,
+            how_to_buy_url,
+            transaction_url_format,
+            standards,
+            payment,
+            now,
+        );
+        info!(name, %ledger_canister_id, "Token added");
+    });
+
+    Success
 }
 
 struct Urls {
@@ -159,4 +220,12 @@ fn extract_urls(
         how_to_buy_url,
         transaction_url_format,
     })
+}
+
+async fn icrc2_transfer_from(ledger_canister_id: CanisterId, transfer_args: &TransferFromArgs) -> Result<u64, String> {
+    match icrc_ledger_canister_c2c_client::icrc2_transfer_from(ledger_canister_id, transfer_args).await {
+        Ok(Ok(block_index)) => Ok(block_index.0.try_into().unwrap()),
+        Ok(Err(err)) => Err(format!("Error calling 'icrc2_transfer_from': {err:?}")),
+        Err(error) => Err(format!("IC error calling 'icrc2_transfer_from': {error:?}")),
+    }
 }