+++ backend/canisters/registry/impl/src/lifecycle/post_upgrade.rs
@@ -25,24 +25,4 @@ fn post_upgrade(args: Args) {
     init_state(env, data, args.wasm_version);
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
-
-    mutate_state(|state| {
-        let now = state.env.now();
-
-        for token in state.data.tokens.iter_mut() {
-            if token.symbol == "CHAT" {
-                token.how_to_buy_url = "https://oc.app/faq?q=buychat".to_string();
-            } else if token.symbol == "ICP" {
-                token.how_to_buy_url = "https://oc.app/faq?q=buyicp".to_string();
-            } else if token.how_to_buy_url == "https://3ezrj-4yaaa-aaaam-abcha-cai.ic0.app/sns/faq#how-can-i-get-sns-tokens" {
-                token.how_to_buy_url = "https://internetcomputer.org/sns/faq#how-can-i-get-sns-tokens".to_string();
-            } else {
-                continue;
-            }
-
-            token.last_updated = now;
-        }
-
-        state.data.tokens.last_updated = now;
-    });
 }